@@ -15,16 +15,17 @@
     "require": {
         "php": ">=5.6",
         "ext-gd": "*",
-        "symfony/options-resolver": "^2.3|^3.0",
+        "symfony/options-resolver": "^3.0",
         "bacon/bacon-qr-code": "^1.0"
     },
     "require-dev": {
-        "symfony/browser-kit": "^2.3|^3.0",
-        "symfony/framework-bundle": "^2.3|^3.0",
-        "symfony/http-kernel": "^2.3|^3.0",
+        "symfony/browser-kit": "^3.0",
+        "symfony/framework-bundle": "^3.0",
+        "symfony/http-kernel": "^3.0",
+        "symfony/yaml": "^3.0",
         "sensio/framework-extra-bundle": "^3.0",
-        "phpunit/phpunit": "^4.0|^5.0",
-        "khanamiryan/qrcode-detector-decoder": "dev-master"
+        "khanamiryan/qrcode-detector-decoder": "dev-master",
+        "phpunit/phpunit": "^5.0|^6.0"
     },
     "autoload": {
         "psr-4": {
